refactor: rename main component, add test case
