+++ src/RowOption/index.js
@@ -1,0 +1,3 @@
+// @flow
+
+export { default as RowOption } from './RowOption';
